pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "marketolog"
version = "0.1.2"
description = "AI-маркетолог для бизнеса в Рунете — MCP-сервер для Claude"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["mcp", "marketing", "seo", "claude", "ai", "runet"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "fastmcp>=3.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
]

[project.scripts]
marketolog = "marketolog.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
